requirements.txt
Pillow>=10.0.0
gspread>=6.0.0
google-auth>=2.0.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
